rename from africa.absa.cps to za.co.absa
