+++ bigfiles/src/main/scala/za/co/absa/Comparator.scala
@@ -14,9 +14,9 @@
  * limitations under the License.
  **/
 
-package africa.absa.cps
+package za.co.absa
 
-import africa.absa.cps.hash.HashUtils.HASH_COLUMN_NAME
+import za.co.absa.hash.HashUtils.HASH_COLUMN_NAME
 import hash.HashUtils
 import org.apache.spark.sql.{DataFrame, SparkSession}
 import org.json4s.JsonAST
